associated refactoring
